requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.57.0
django-apscheduler==0.6.2
django-ckeditor==6.7.0
django-filter==23.3
django-js-asset==2.1.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.0.7
